introduce unit tests for leiningen.search/parse
This commit adds a new namespace, leiningen.test.search, where we set
up an HTTP server mocking Clojars error responses when invalid queries
are sent. This is done to test the behavior of leiningen.search/parse
when a 400 repsonse code is returned from the server. The overall
result is ensuring the previous commit improves error reporting in the
search task.