Alway use `[[ ...  ]]` conditionals
The `[` is just a command with args and expects/ignores the `]` final arg.

That means you need to be very concerned about what's inside `[ ... ]`.

The `[[ ... ]]` is Bash syntax and everything inside is handled as
syntax.
That means you don't need to quotes variable expansions.
Bash knows what you want and does it properly.

Some related changes were made in this context:

* [[ $foo ]] same as [[ -n $foo ]] same as [[ $foo != "" ]]
* [[ -z $foo ]] same as [[ $foo == "" ]]
* Prefer == over = for string equality
* Prefer -eq over == for numeric equality
* $* means all aregs expanded into 1 string
* Words without meta chars are strings in Bash. No need to quote them.

Note: the RHS of a == or != op must be quoted if it is a variable
expansion, but shellcheck would complain if you didn't quote it.

All the files are still shellcheck compliant.

Very confident of this commit.