AOT compile all namespaces required by leiningen.core.main
We have some namespaces that are always required, regardless of which
task is running. We were not AOT compiling a few of these namespaces,
which caused a small slowdown in startup for any task (because the
Clojure compile would need to emit bytecode at runtime for all these
namespaces).

Simply including these namespaces in the AOT vector reduces over 200ms
of execution time in simple tasks:

```
$ hyperfine 'lein version'
Benchmark 1: lein version
  Time (mean ± σ):     646.6 ms ±   4.5 ms    [User: 805.5 ms, System: 58.8 ms]
  Range (min … max):   641.3 ms … 656.0 ms    10 runs

$ hyperfine 'lein-main version'
Benchmark 1: lein-main version
  Time (mean ± σ):     413.8 ms ±   3.3 ms    [User: 461.8 ms, System: 47.0 ms]
  Range (min … max):   408.6 ms … 417.8 ms    10 runs
```