Offer opt-in to include metadata in code to be evaluated in project
Add a new project key :preserve-eval-meta, to determine if metadata should be included in the code that will be evaluated in project.

Enabling that flag prevents reflection warnings when running `lein run` in a project with `:global-vars {*warn-on-reflection* true}`, but it may break plugins that spit code including objects with metadata that can't be read.

Fixes https://github.com/technomancy/leiningen/issues/2814.
Fixes https://github.com/technomancy/leiningen/issues/2695.