Ensure :active-profiles respects unmerging, say via lein pom/java
Although a project's :active-profiles metadata is not currently part
of the publically specified API, it can be useful in middlewares.

Previously init-profiles, which is used by unmerge-profiles carried
over the :active-profiles from the incoming project, and since
:active-profiles is built solely via conj, this ended up not removing
profiles that were being unmerged.  To fix that, just clear the active
profiles in init-profiles before it rebuilds them.